requirements.txt
fastapi
uvicorn
bcrypt
pymongo
python-dotenv
python-jose
pytz
pydantic[email]
aiohttp